+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/services/UIMetadataService.java
@@ -20,8 +20,11 @@ import ca.corefacility.bioinformatics.irida.model.sample.metadata.MetadataRestri
 import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ui.SelectOption;
 import ca.corefacility.bioinformatics.irida.ria.web.ajax.metadata.dto.ProjectMetadataField;
 import ca.corefacility.bioinformatics.irida.ria.web.ajax.metadata.dto.ProjectMetadataTemplate;
+import ca.corefacility.bioinformatics.irida.ria.web.projects.settings.dto.Role;
 import ca.corefacility.bioinformatics.irida.service.ProjectService;
 import ca.corefacility.bioinformatics.irida.service.sample.MetadataTemplateService;
+
+import com.google.common.collect.ImmutableList;
 
 /**
  * Service for Metadata Templates in the user interface
@@ -29,6 +32,7 @@ import ca.corefacility.bioinformatics.irida.service.sample.MetadataTemplateServi
 @Component
 public class UIMetadataService {
 	private static final Logger logger = LoggerFactory.getLogger(UIMetadataService.class);
+	private final List<String> METADATA_ROLES = ImmutableList.of("LEVEL_1", "LEVEL_2", "LEVEL_3", "LEVEL_4");
 
 	private final ProjectService projectService;
 	private final MetadataTemplateService templateService;
@@ -201,6 +205,18 @@ public class UIMetadataService {
 			List<MetadataTemplateField> fields) {
 		return fields.stream()
 				.map(field -> createProjectMetadataField(project, field))
+				.collect(Collectors.toList());
+	}
+
+	/**
+	 * Get a list of all metadata roles
+	 *
+	 * @param locale current users {@link Locale}
+	 * @returnList of metadata roles that are available to the suer
+	 */
+	public List<Role> getProjectMetadataRoles(Locale locale) {
+		return METADATA_ROLES.stream()
+				.map(role -> new Role(role, messageSource.getMessage("metadataRole." + role, new Object[] {}, locale)))
 				.collect(Collectors.toList());
 	}
